@@ -151,14 +151,14 @@
         <div class="container-fluid">
             <div class="row">
                 <div class="col-12 col-md-6">
-                    <div class="row">
+                    <div class="row mb-3">
                         <div class="col">
                             <a asp-controller="Home" asp-action="Manual">
                                 <i class="fas fa-info-circle fa-tw"></i> @Localizer["User Manual"]
                             </a>
                         </div>
                     </div>
-                    <div class="row">
+                    <div class="row mb-3">
                         <div class="col">
                             <a asp-controller="Home" asp-action="Privacy">
                                 <i class="fas fa-info-circle fa-lock"></i> @Localizer["Privacy Policy"]
@@ -167,7 +167,7 @@
                     </div>
                 </div>
                 <div class="col-12 col-md-6">
-                    <a target="_blank" href="https://www.epicsolutions.cl" class="float-right" rel="noopener">
+                    <a target="_blank" href="https://www.epicsolutions.cl" class="float-left-md" rel="noopener">
                         <img src="~/epic_light_hor.svg" alt="epic solutions" class="img-responsive" height="20" asp-append-version="true" /> &copy; @DateTime.Now.Year - BiblioMit
                     </a>
                 </div>
